Welcome Module
View: Create Folder in templates/welcome and file welcome.html here we have the welcome page of the app.
SASS: In folder scss create a folder for the welcome module view, there you will find the css styles for the view welcome.html.
Route: in app.js create the .state for the route and link the template.
Next: Login and Register.